requirements.txt
dnspython
stem
pysocks
